pyproject.toml
[tool.poetry]
name = "university-bot"
version = "0.1.0"
description = ""
authors = ["ivalukyan <ivalukyanov@gmail.com>"]
license = "LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.13.0"
python-dotenv = "^1.0.1"
pytz = "^2024.1"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.34"
pandas = "^2.2.2"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
